testing is done

+++ src/main/java/com/cts/employee/controller/EmployeeController.java
@@ -4,12 +4,10 @@ import com.cts.employee.model.Employee;
 import com.cts.employee.repo.EmployeeRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/employees")
@@ -52,6 +50,10 @@ public class EmployeeController {
 
     @DeleteMapping("/{id}")
     void deleteEmployee(@PathVariable Long id) {
-        repo.deleteById(id);
+        if (repo.existsById(id)) {
+            repo.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 }
